Remove previous notification model
